refactor(summary): split evidence reserve policy
